#CLOUD-328 one more unit test
#VMS-2270 fixed: 'find system' multicast packet did not contain cloudSystemId. That resulted in very long cloud system discovery in LAN
#VMS-2391 CloudServerSocket sends listen request after connection to mediator has been restored
#VMS-2391 RetryTimer: added reset to inital state for re-usage
#VMS-2391 added declaration of subscription to stun client tcp reconnect event
#VMS-2391 fixed compile error
#vms-2401. Archive camera info file is created on per storage basis rather than per storage pool.
#vms-2401. Archive camera info file is created on per storage basis rather than per storage pool. Merge heads.
CLOUD-324 network::cloud::AddressResolver fix, tests and refactoring
VMS-2154 Put all NICs up before calculating HardwareID on Windows
address review commments
fixed compile error
fixed compile warnings
minor compile fix
some review fixes

<==== Mercurial data ====>
Branch: prod_3.0.0

+++ common_libs/nx_network/src/nx/network/connection_server/base_stream_protocol_connection.h
@@ -69,7 +69,7 @@ namespace nx_api
 
         void bytesReceived( const nx::Buffer& buf )
         {
-            for( size_t pos = 0; pos < buf.size(); )
+            for( size_t pos = 0; pos < (size_t)buf.size(); )
             {
                 //parsing message
                 size_t bytesProcessed = 0;
@@ -77,7 +77,7 @@ namespace nx_api
                 {
                     case ParserState::init:
                     case ParserState::inProgress:
-                        NX_ASSERT( bytesProcessed == buf.size() );
+                        NX_ASSERT( bytesProcessed == (size_t)buf.size() );
                         return;
 
                     case ParserState::done:
